Warnings in debug

+++ dali/kernels/tensor_shape.h
@@ -513,7 +513,7 @@ struct TensorListShapeBase {
   template <typename SampleShape>
   void set_tensor_shape(int64_t sample, const SampleShape &sample_shape) {
     detail::check_compatible_ndim<sample_ndim, compile_time_size<SampleShape>::value>();
-    assert(dali::size(sample_shape) == sample_dim());
+    assert(static_cast<int>(dali::size(sample_shape)) == static_cast<int>(sample_dim()));
     int64_t base = sample_dim() * sample;
     for (int i = 0; i < sample_dim(); i++) {
       shapes[base + i] = sample_shape[i];
